+++ docs/history.md
@@ -95,6 +95,8 @@
 ||⚠️BREAKING CHANGE: Remove support of `colors`|
 ||⚠️BREAKING CHANGE: Remove support of `mime`|
 ||⚠️BREAKING CHANGE: Remove option `watch` in `custom`|
+||⚠️BREAKING CHANGE: Behavior of `configuration` in `custom`|
+||⚠️BREAKING CHANGE: Behavior of capturing groups in `custom`|
 ||⚠️BREAKING CHANGE: Remove option `strict` in `file`|
 ||⚠️BREAKING CHANGE: Remove option `case-sensitive` in `file`, it is now always case sensitive|
 ||⚠️BREAKING CHANGE: Remove option `ignore-if-not-found` in `file`, mapping is ignored if resolving to a missing / unreadable / invalid path|