+++ docs/history.md
@@ -112,6 +112,7 @@
 ||⚠️BREAKING CHANGE: Mocked Request normalizes the received URL|
 ||⚠️BREAKING CHANGE: `body` helper leverages request headers and offers conversion interface|
 ||⚠️BREAKING CHANGE: `server-created` event is now renamed to `created`|
+||⚠️BREAKING CHANGE: `match` interpretation and interpolation has evolved|
 ||Fix `use` to handle middleware that are not calling next|
 ||Response status code is preserved when set before `file`|
 ||`static` option for performant `file` delivery ([`doc`](file.md))|